+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Piral Changelog
 
+## 1.8.5 (tbd)
+
+- Fixed generation of declarations during `pilet build` in the `piral-cli`
+- Updated versions of Angular used in `piral-ng`
+- Added support for Angular 19 in `piral-ng` incl. support for input signals
+
 ## 1.8.4 (April 1, 2025)
 
 - Fixed newline when updating JSON files using the `piral-cli` (#749)